Fix missing refactorings
